backend/requirements.txt
asgiref==3.10.0
certifi==2025.10.5
Django==5.2.7
django-cors-headers==4.9.0
django-ratelimit==4.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-extensions==0.8.0
packaging==25.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-decouple==3.8
sentry-sdk==2.43.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
